TTS: share the row-selection dialog scaffolding
The detect-speakers and skip-noise-lines dialogs were near-verbatim
copies of the same checkbox-review dialog. A shared row base, view-model
base, and window builder now hold the common structure (grid, table with
checkbox column, Ok/Cancel/SelectAll/Invert, Escape and space handling,
window persistence); each dialog keeps only its own columns, panels, and
result collection. The unused SelectedRow property is gone from both
view models.
