+++ src/ui/Features/Video/TextToSpeech/SkipNoiseLines/SkipNoiseLineRow.cs
@@ -1,25 +1,16 @@
-using CommunityToolkit.Mvvm.ComponentModel;
 using Nikse.SubtitleEdit.Core.Common;
 
 namespace Nikse.SubtitleEdit.Features.Video.TextToSpeech.SkipNoiseLines;
 
-public partial class SkipNoiseLineRow : ObservableObject
+public class SkipNoiseLineRow : SelectLinesRowBase
 {
-    [ObservableProperty] private bool _isSelected;
-
-    public int Number { get; }
-    public string Show { get; }
-    public string Text { get; }
     public Paragraph Paragraph { get; }
 
     public SkipNoiseLineRow(Paragraph paragraph)
-    {
         // Every detected line starts checked - the whole point of the dialog is skipping them,
         // unchecking is for the occasional false positive.
-        IsSelected = true;
-        Number = paragraph.Number;
-        Show = paragraph.StartTime.ToDisplayString();
-        Text = paragraph.Text;
+        : base(true, paragraph.Number, paragraph.StartTime.ToDisplayString(), paragraph.Text)
+    {
         Paragraph = paragraph;
     }
 }
